Improved project creation #68
The context properties can now be set by the individual components of the ui,
allowing each project to access different instances of the components.

As object creation for a new tab of the qml TabView can not be controlled (and
the context therefore not customized), an empty item is loaded by the TabView first.
When this item is completed, a c++ method is invoked and the new project is
created in a context with the required context properties.